[pull] dev from NuGet:dev

+++ src/NuGet.Core/NuGet.Packaging/PackageExtractor.cs
@@ -398,12 +398,6 @@ namespace NuGet.Packaging
                             logger.LogVerbose(
                                 $"Acquired lock for the installation of {packageIdentity.Id} {packageIdentity.Version}");
 
-                            logger.LogInformation(string.Format(
-                                CultureInfo.CurrentCulture,
-                                Strings.Log_InstallingPackage,
-                                packageIdentity.Id,
-                                packageIdentity.Version));
-
                             cancellationToken.ThrowIfCancellationRequested();
 
                             // We do not stop the package extraction after this point
@@ -432,6 +426,7 @@ namespace NuGet.Packaging
                             var tempHashPath = Path.Combine(targetPath, Path.GetRandomFileName());
                             var tempNupkgMetadataPath = Path.Combine(targetPath, Path.GetRandomFileName());
                             var packageSaveMode = packageExtractionContext.PackageSaveMode;
+                            string contentHash;
 
                             try
                             {
@@ -494,12 +489,13 @@ namespace NuGet.Packaging
                                         File.WriteAllText(tempHashPath, packageHash);
 
                                         // get hash for the unsigned content of package
-                                        var contentHash = packageReader.GetContentHash(cancellationToken, GetUnsignedPackageHash: () => packageHash);
+                                        contentHash = packageReader.GetContentHash(cancellationToken, GetUnsignedPackageHash: () => packageHash);
 
                                         // write the new hash file
                                         var hashFile = new NupkgMetadataFile()
                                         {
-                                            ContentHash = contentHash
+                                            ContentHash = contentHash,
+                                            Source = source
                                         };
 
                                         NupkgMetadataFileFormat.Write(tempNupkgMetadataPath, hashFile);
@@ -553,7 +549,7 @@ namespace NuGet.Packaging
 
                             File.Move(tempNupkgMetadataPath, nupkgMetadataFilePath);
 
-                            logger.LogVerbose($"Completed installation of {packageIdentity.Id} {packageIdentity.Version}");
+                            logger.LogInformation(StringFormatter.Log_InstalledPackage(packageIdentity.Id, packageIdentity.Version.OriginalVersion, source, contentHash, targetPath));
 
                             packageExtractionTelemetryEvent.SetResult(NuGetOperationStatus.Succeeded);
                             return true;
@@ -588,8 +584,8 @@ namespace NuGet.Packaging
 
                 // delete the parent directory if it is empty
                 if (Directory.Exists(parent.FullName) &&
-                parent.GetFiles().Count() == 0 &&
-                parent.GetDirectories().Count() == 0)
+                    !parent.GetFiles().Any() &&
+                    !parent.GetDirectories().Any())
                 {
                     Directory.Delete(parent.FullName);
                 }
@@ -640,12 +636,6 @@ namespace NuGet.Packaging
                         {
                             logger.LogVerbose(
                                 $"Acquired lock for the installation of {packageIdentity.Id} {packageIdentity.Version}");
-
-                            logger.LogInformation(string.Format(
-                                CultureInfo.CurrentCulture,
-                                Strings.Log_InstallingPackage,
-                                packageIdentity.Id,
-                                packageIdentity.Version));
 
                             cancellationToken.ThrowIfCancellationRequested();
 
@@ -786,7 +776,8 @@ namespace NuGet.Packaging
                             // write the new hash file
                             var hashFile = new NupkgMetadataFile()
                             {
-                                ContentHash = contentHash
+                                ContentHash = contentHash,
+                                Source = packageDownloader.Source
                             };
 
                             NupkgMetadataFileFormat.Write(tempNupkgMetadataFilePath, hashFile);
@@ -825,7 +816,7 @@ namespace NuGet.Packaging
 
                             File.Move(tempNupkgMetadataFilePath, nupkgMetadataFilePath);
 
-                            logger.LogVerbose($"Completed installation of {packageIdentity.Id} {packageIdentity.Version}");
+                            logger.LogInformation(StringFormatter.Log_InstalledPackage(packageIdentity.Id, packageIdentity.Version.OriginalVersion, packageDownloader.Source, contentHash, targetPath));
 
                             packageExtractionTelemetryEvent.SetResult(NuGetOperationStatus.Succeeded);
                             return true;
@@ -1068,7 +1059,7 @@ namespace NuGet.Packaging
                         if (packageExtractionContext.Logger is ICollectorLogger collectorLogger)
                         {
                             // collectorLogger.Errors is a collection of errors and warnings, we just need to fail if there are errors.
-                            if (collectorLogger.Errors.Where(e => e.Level >= LogLevel.Error).Any())
+                            if (collectorLogger.Errors.Any(e => e.Level >= LogLevel.Error))
                             {
                                 // Send empty results since errors and warnings have already been logged
                                 throw new SignatureException(results: Enumerable.Empty<PackageVerificationResult>().ToList(), package: package);
